Finished testing suite.
Created a python script that automates all the tests for all the maps.
Stores result in a text file.

So apparently the bot thread (part of the original code) doesn't
wanna get interrupted even after we call .interrupt(). The only
way to kill it was by killing the main program. Thus, having Python
spawn a different process for each test was the only viable solution
(as opposed to having a Java class instatiate them manually).

The initially idea involved the latter solution, and that ended terribly...
the spwaned threads were never collected and the program got laggier with time.